--- a/python/sglang/srt/managers/scheduler_output_processor_mixin.py
+++ b/python/sglang/srt/managers/scheduler_output_processor_mixin.py
@@ -54,7 +54,7 @@ class SchedulerOutputProcessorMixin:
 
             if self.enable_overlap:
                 logits_output, next_token_ids, _ = (
-                    self.tp_worker.resolve_last_batch_result(launch_done)
+                    self.tp_worker.resolve_last_batch_result(None)
                 )
             else:
                 # Move next_token_ids and logprobs to cpu
@@ -196,6 +196,8 @@ class SchedulerOutputProcessorMixin:
                     req.is_chunked -= 1
 
         self.stream_output(batch.reqs, batch.return_logprob, skip_stream_req)
+        if self.is_generation and self.enable_overlap and launch_done is not None:
+            launch_done.wait()
 
     def process_batch_result_decode(
         self: Scheduler,
@@ -212,7 +214,7 @@ class SchedulerOutputProcessorMixin:
 
         if self.enable_overlap:
             logits_output, next_token_ids, can_run_cuda_graph = (
-                self.tp_worker.resolve_last_batch_result(launch_done)
+                self.tp_worker.resolve_last_batch_result(None)
             )
             next_token_logprobs = logits_output.next_token_logprobs
         elif batch.spec_algorithm.is_none():
@@ -292,6 +294,8 @@ class SchedulerOutputProcessorMixin:
 
         self.set_next_batch_sampling_info_done(batch)
         self.stream_output(batch.reqs, batch.return_logprob)
+        if self.enable_overlap and launch_done is not None:
+            launch_done.wait()
         self.token_to_kv_pool_allocator.free_group_end()
 
         self.forward_ct_decode = (self.forward_ct_decode + 1) % (1 << 30)