<!-- Thank you for your contribution! Please follow these guidelines to enhance your pull request. If anything is unclear, submit your PR and reach out to maintainers for assistance. Join our Slack community at https://slack.sglang.ai to discuss further. -->

## Motivation

When the overlap schedule is enabled, resolve_last_batch_result first waits for the launch_done of the current batch and the copy_done of the previous batch, then processes the output of the previous batch. However, if the forward process of the current batch is not fast enough, the output processing of the previous batch will have to wait further.

 **_We can swap the waiting order of launch_done and copy_done, so that once launch_done is ready, the output processing can begin, and after the output processing is completed, we continue waiting for copy_done_**. 

Since process_batch_result hands over the necessary output processing to the DetokenizerManager after some preliminary processing, it will not compete too much with the forward thread for CPU resources.

<!-- Describe the purpose and goals of this pull request. -->

## Modifications

Modify the waiting order of launch_done and copy_done in resolve_last_batch_result. First, wait for copy_done, and once the condition is met, begin processing the output of the previous batch, then wait for the launch_done of the current batch.

<!-- Detail the changes made in this pull request. -->

## Accuracy Tests

<!-- If this pull request affects model outputs (e.g., changes to the kernel or model forward code), provide accuracy test results. -->

`python3 -m sglang.test.few_shot_gsm8k --num-questions 200`

> Accuracy: 0.980
Invalid: 0.000
Latency: 28.535 s
Output throughput: 987.985 token/s
## Benchmarking and Profiling

<!-- If this pull request impacts inference speed, provide benchmarking and profiling results. -->

|  | Before | After | Performance gain |
| --- | --- | --- | --- |
| P99 TTFT (ms) | 3413.01 | 2188.50 |**-35.8%** |
| P99 ITL (ms) | 27.28 | 27.20 |-0.3% |
| Mean E2E Latency (ms) | 41976.97 | 41920.56 |-0.1%|


Server command:

`python3 -m sglang.launch_server --model-path Qwen/Qwen3-235B-A22B-Instruct-2507-FP8 --enable-metrics 
        --disable-radix-cache 
        --mem-fraction-static 0.65 
        --tp-size 8 
        --ep-size 8 
        --enable-dp-attention --dp-size 8 
        --chunked-prefill-size -1 
        --enable-deepep-moe 
        --cuda-graph-max-bs 128 `

Client command:

`python -m sglang.bench_serving --backend sglang --dataset-name random --dataset-path /mnt/ShareGPT_V3_unfiltered_cleaned_split.json --random-input-len 3500 --random-output-len 1500 --request-rate 4   --num-prompts 20  --random-range-ratio 1`

Before

> ============ Serving Benchmark Result ============
Backend:                                 sglang
Traffic request rate:                    4.0
Max request concurrency:                 not set
Successful requests:                     20
Benchmark duration (s):                  43.27
Total input tokens:                      70000
Total generated tokens:                  30000
Total generated tokens (retokenized):    30000
Request throughput (req/s):              0.46
Input token throughput (tok/s):          1617.73
Output token throughput (tok/s):         693.31
Total token throughput (tok/s):          2311.04
Concurrency:                             19.40
----------------End-to-End Latency----------------
Mean E2E Latency (ms):                   41976.97
Median E2E Latency (ms):                 42249.84
---------------Time to First Token----------------
Mean TTFT (ms):                          2331.79
Median TTFT (ms):                        2407.76
P99 TTFT (ms):                           3413.01
---------------Inter-Token Latency----------------
Mean ITL (ms):                           26.45
Median ITL (ms):                         26.08
P95 ITL (ms):                            26.88
P99 ITL (ms):                            27.28
Max ITL (ms):                            1282.86
==================================================

After:

>============ Serving Benchmark Result ============
Backend:                                 sglang
Traffic request rate:                    4.0
Max request concurrency:                 not set
Successful requests:                     20
Benchmark duration (s):                  43.21
Total input tokens:                      70000
Total generated tokens:                  30000
Total generated tokens (retokenized):    30000
Request throughput (req/s):              0.46
Input token throughput (tok/s):          1619.86
Output token throughput (tok/s):         694.22
Total token throughput (tok/s):          2314.08
Concurrency:                             19.40
----------------End-to-End Latency----------------
Mean E2E Latency (ms):                   41920.56
Median E2E Latency (ms):                 42194.16
---------------Time to First Token----------------
Mean TTFT (ms):                          1507.03
Median TTFT (ms):                        1416.36
P99 TTFT (ms):                           2188.50
---------------Inter-Token Latency----------------
Mean ITL (ms):                           26.96
Median ITL (ms):                         26.15
P95 ITL (ms):                            26.89
P99 ITL (ms):                            27.20
Max ITL (ms):                            2276.60
==================================================

## Checklist

- [x] Format your code according to the [Format code with pre-commit](https://docs.sglang.ai/developer_guide/contribution_guide.html#format-code-with-pre-commit).
- [ ] Add unit tests according to the [Run and add unit tests](https://docs.sglang.ai/developer_guide/contribution_guide.html#run-and-add-unit-tests).
- [ ] Update documentation according to [Write documentations](https://docs.sglang.ai/developer_guide/contribution_guide.html#write-documentations).
- [x] Provide accuracy and speed benchmark results according to [Test the accuracy](https://docs.sglang.ai/developer_guide/contribution_guide.html#test-the-accuracy) and [Benchmark the speed](https://docs.sglang.ai/developer_guide/contribution_guide.html#benchmark-the-speed).